requirements.txt
# Browser automation
playwright>=1.40.0
selenium>=4.15.0

# LLM integration for natural language processing
openai>=1.3.0
anthropic>=0.7.0

# Data handling and validation
pydantic>=2.5.0
pyyaml>=6.0.1
jsonschema>=4.20.0

# Testing and utilities
pytest>=7.4.0
pytest-asyncio>=0.21.0
pillow>=10.1.0

# CLI and logging
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0
